requirements.txt
Django==3.2.7
psycopg2-binary==2.9.3
djangorestframework==3.13.1
drf-yasg==1.20.0
faker